requirements.txt
requests
mailchimp_marketing
python-dotenv